+++ editoragibi/src/main/java/com/grupo2/editoragibi/Data/Repositories/Interfaces/IGibiRepository.java
@@ -11,11 +11,12 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IGibiRepository extends JpaRepository<GibiEntity, Long>{
+public interface IGibiRepository extends JpaRepository<GibiEntity, Integer>{
 
     @Query("SELECT g FROM GibiEntity g WHERE g.tituloGibi = ?1")
     Optional<GibiEntity> findGibiByTitulo(String titulo);
 
     @Query("SELECT e FROM EdicoesGibiEntity e WHERE e.edicaoGibi_id = ?1")
-    Optional<EdicoesGibiEntity> findEdicaoGibiById(Long id);
+    Optional<EdicoesGibiEntity> findEdicaoGibiById(Integer id);
+
 }
